forgot this line

+++ src/cores/MCD212/DRAMInterface.cpp
@@ -13,7 +13,7 @@ uint8_t MCD212::GetByte(const uint32_t addr, const uint8_t flags)
     if(addr <= 0x4FFFFF)
     {
         const uint8_t data = internalRegisters[addr-0x4FFFE0] & 0x00FF;
-        LOG(if(flags & Log) { out_dram << std::setw(6) << std::hex << app->cpu->currentPC << " Get register at 0x" << std::setw(6) << std::setfill('0') << addr << " : 0x" << data << std::endl; })
+        LOG(if(flags & Log) { out_dram << std::setw(6) << std::hex << app->cpu->currentPC << " Get register at 0x" << std::setw(6) << std::setfill('0') << addr << " : 0x" << (int)data << std::endl; })
         if(addr == 0x4FFFE1 && flags & Trigger)
         {
             internalRegisters[CSR2R] &= 0x00FE; // clear BE bit on status read
